+++ tests/test_config.py
@@ -3,7 +3,7 @@
 from pathlib import Path
 from unittest import TestCase
 
-from baygon.suite import TestSuite
+from baygon.suite import SuiteLoader
 
 
 dir_path = Path(__file__).resolve(strict=True).parent
@@ -11,7 +11,9 @@ dir_path = Path(__file__).resolve(strict=True).parent
 
 class TestConfig(TestCase):
     def test_config_file(self):
-        ts = TestSuite(path="tests")
-        self.assertIn("t.yml", str(ts.path))
-        self.assertIn("version", ts.config)
-        self.assertEqual(ts.config["version"], 1)
+        loader = SuiteLoader()
+        context = loader.load(path="tests")
+
+        self.assertIn("t.yml", str(context.source_path))
+        self.assertIn("version", context.config)
+        self.assertEqual(context.config["version"], 1)